fix for upgradation to julia 1.6.0 (#14)
* use Int instead of Int64 in Room

* fix DoorKey: use proper types for partitions

* fix Makie.jl related bugs for graphical rendering

* update Project.toml for benchmarking

+++ src/envs/envs.jl
@@ -1,5 +1,5 @@
 struct Room
-    region::CartesianIndices{2,Tuple{UnitRange{Int64},UnitRange{Int64}}}
+    region::CartesianIndices{2,Tuple{UnitRange{Int},UnitRange{Int}}}
 end
 
 Room(origin, height, width) = Room(CartesianIndices((origin.I[1] : origin.I[1] + height - 1, origin.I[2] : origin.I[2] + width - 1)))
